+++ examples/ExampleFeatures1.hs
@@ -4,49 +4,33 @@ Description : Demostrates how to define features using Hasklepias
 Copyright   : (c) NoviSci, Inc 2020
 License     : BSD3
 Maintainer  : bsaul@novisci.com
-Stability   : experimental
 -}
 
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE FlexibleContexts #-}
-
+{-# LANGUAGE MonoLocalBinds #-}
 module ExampleFeatures1(
     exampleFeatures1Spec
-  , dropEventFilter
-  , index
-  , baseline
-  , enrolled
-  , enrolled'
-  , pack
 ) where
 
 import Hasklepias
 import ExampleEvents
 import Test.Hspec
-import Data.Text(pack, Text)
+import Data.Text(Text)
 import Data.Maybe ( fromMaybe )
--- import Control.Monad
-
-{- Helper functions used below -}
-
-dropEventFilter :: (IntervalAlgebraic Interval a)=>
-     (Events a -> Events a)
-  -> [Interval a]
-  -> [Interval a]
-dropEventFilter f l = (intervals . f) (fmap (`event` emptyContext) l)
 
 {-
 Index is defined as the first occurrence of an Orca bite.
 -}
-index:: (IntervalAlgebraic Interval a) =>
-     Events a
-  -> Feature (Interval a)
-index es =
-    case firstConceptOccurrence ["wasBitByOrca"] es of
-        Nothing -> Left $ Other "No occurrence of Orca bite"
-        Just x  -> Right (getInterval x)
+
+indexDef :: (IntervalAlgebraic Interval a) =>
+          FeatureDefinition e a (Interval a)
+indexDef = defineEF
+      (Other "No occurrence of Orca bite")
+      (firstConceptOccurrence ["wasBitByOrca"])
+      getInterval
 
 {-  
 The baseline interval is the interval (b - 60, b), where b is the begin of 
@@ -62,137 +46,101 @@ baseline = fmap (enderval 60 . begin)
 bline :: (IntervalAlgebraic Interval a, IntervalSizeable a b) =>
      Events a
   -> Feature (Interval a)
-bline = baseline.index
+bline = baseline . applyEF indexDef
 
 flwup :: (IntervalAlgebraic Interval a, IntervalSizeable a b) =>
      Events a
   -> Feature (Interval a)
-flwup = fmap (beginerval 30 . begin) . index
+flwup = fmap (beginerval 30 . begin) . applyEF indexDef
+
 {-
 Define enrolled as the indicator of whether all of the gaps between the union of 
 all enrollment intervals (+ allowableGap) 
 -}
-enrolled :: (IntervalSizeable a b, IntervalCombinable a) =>
-      Feature (Interval a)
-   -> Events a
-   -> Feature Bool
-enrolled (Left  _) _ = Left Excluded
-enrolled (Right i) l = Right $ enrolled' 8 i l
-
-enrolled' :: (IntervalSizeable a b, IntervalCombinable a) =>
-      b  -- ^ allowable gap between enrollment intervals
-   -> Interval a -- ^ baseline interval
-   -> Events a
-   -> Bool
-enrolled' allowableGap i =
+enrolledDef :: IntervalSizeable a b => b -> FeatureDefinition (Interval a) a Bool
+enrolledDef allowableGap = defineFEF Excluded
+   (  \i ->
         maybe False (all (< allowableGap) . durations)
       . gapsWithin i
       . combineIntervals
       . intervals
       . makeConceptsFilter ["enrollment"]
+   )
 
 {-
 Define features that identify whether a subject as bit/struck by a duck and
 bit/struck by a macaw.
 -}
-makeHasAnyHistoryFeature :: (IntervalAlgebraic Interval a
-                             , IntervalAlgebraic (PairedInterval Context) a) =>
-       [Text]
-    -> Feature (Interval a)
-    -> Events a
-    -> Feature (Bool, Maybe (Interval a))
-makeHasAnyHistoryFeature cnpts feat es =
-  case feat of
-   (Left x ) -> Left Excluded
-   (Right i) -> Right (isNotEmpty candidates, safeLast $ intervals candidates)
-      where candidates = makePairedFilter enclose i (`hasConcepts` cnpts) es
+makeHxDef :: (IntervalAlgebraic Interval a) =>
+               [Text] -> FeatureDefinition (Interval a) a (Bool, Maybe (Interval a))
+makeHxDef cnpts = defineFEF Excluded
+   ( \i es ->
+      (isNotEmpty (f i es), safeLast $ intervals (f i es))
+   )
+   where f i x = makePairedFilter enclose i (`hasConcepts` cnpts) x
 
-hasDuckHistory :: (IntervalAlgebraic Interval a
-                  , IntervalSizeable a b
-                  , IntervalAlgebraic (PairedInterval Context) a) =>
-       Events a
-    -> Feature (Bool, Maybe (Interval a))
-hasDuckHistory x = makeHasAnyHistoryFeature
-   ["wasBitByDuck", "wasStruckByDuck"]
-   (bline x ) x
+duckHxDef :: (IntervalAlgebraic Interval a) =>
+          FeatureDefinition (Interval a) a (Bool, Maybe (Interval a))
+duckHxDef = makeHxDef ["wasBitByDuck", "wasStruckByDuck"]
 
-hasMacawHistory :: (IntervalAlgebraic Interval a
-                  , IntervalSizeable a b
-                  , IntervalAlgebraic (PairedInterval Context) a) =>
-      Events a
-    -> Feature (Bool, Maybe (Interval a))
-hasMacawHistory x = makeHasAnyHistoryFeature
-   ["wasBitByMacaw", "wasStruckByMacaw"]
-   (bline x) x
+macawHxDef :: (IntervalAlgebraic Interval a) =>
+          FeatureDefinition (Interval a) a (Bool, Maybe (Interval a))
+macawHxDef = makeHxDef ["wasBitByMacaw", "wasStruckByMacaw"]
 
-{-
-Define an event that identifies whether the subject has two minor or one major
-surgery 
--}
-
-twoMinorOrOneMajor :: (IntervalAlgebraic Interval a
-                      , IntervalAlgebraic (PairedInterval Context) a) =>
-     Feature (Interval a)
-  -> Events a
-  -> Feature Bool
-twoMinorOrOneMajor feat l =
- case feat of
-   Left x  -> Left Excluded
-   Right i -> Right $ twoXOrOneY ["hadMinorSurgery"] ["hadMajorSurgery"] (filterEnclose i l)
-
+-- | Define an event that identifies whether the subject has two minor or one major
+--   surgery.
+twoMinorOrOneMajorDef :: (IntervalAlgebraic Interval a) =>
+         FeatureDefinition (Interval a) a Bool
+twoMinorOrOneMajorDef = defineFEF Excluded
+      ( \i es ->
+          twoXOrOneY ["hadMinorSurgery"] ["hadMajorSurgery"] (filterEnclose i es)
+      )
 
 -- | Time from end of baseline to end of most recent Antibiotics
 --   with 5 day grace period
-timeSinceLastAntibiotics :: (IntervalAlgebraic Interval a, IntervalCombinable a, IntervalSizeable a b) =>
-         Feature (Interval a)
-      -> Events a
-      -> Feature (Maybe b)
-timeSinceLastAntibiotics (Left _) _   = Left Excluded
-timeSinceLastAntibiotics (Right i) es = Right $ timeSinceLastAntibiotics' i es
-
-
-timeSinceLastAntibiotics' :: (IntervalAlgebraic Interval a, IntervalCombinable a, IntervalSizeable a b) =>
-      Interval a -> Events a -> Maybe b
-timeSinceLastAntibiotics' blinterval =
-   safeLast                                   -- want the last one
-  .map (max 0 . diff (end blinterval) . end) -- distances between end of baseline and antibiotic intervals
-  .filterNotDisjoint blinterval           -- filter to intervals not disjoint from baseline interval
-  .combineIntervals                       -- combine overlapping intervals
-  .map (expandr 5)                        -- allow grace period
-  .intervals                              -- extract intervals
-  .makeConceptsFilter ["tookAntibiotics"]  -- filter to only antibiotics events
-
+timeSinceLastAntibioticsDef :: ( IntervalAlgebraic Interval a
+                               , IntervalSizeable a b) =>
+         FeatureDefinition (Interval a) a (Maybe b)
+timeSinceLastAntibioticsDef = defineFEF Excluded
+      ( \i es ->
+           ( safeLast                                 -- want the last one
+           . map (max 0 . diff (end i) . end)       -- distances between end of baseline and antibiotic intervals
+           . filterNotDisjoint i                    -- filter to intervals not disjoint from baseline interval
+           . combineIntervals                       -- combine overlapping intervals
+           . map (expandr 5)                        -- allow grace period
+           . intervals                              -- extract intervals
+           . makeConceptsFilter ["tookAntibiotics"])-- filter to only antibiotics events 
+            es
+      )
 
 -- | Count of hospital events in a interval and duration of the last one
-countOfHospitalEvents :: (IntervalCombinable a, IntervalSizeable a b) =>
-     Feature (Interval a)
-  -> Events a
-  -> Feature (Int, Maybe b)
-countOfHospitalEvents (Left _) _   = Left Excluded
-countOfHospitalEvents (Right i) es = Right $
+countOfHospitalEventsDef :: (IntervalCombinable Interval a
+                            , IntervalSizeable a b) =>
+                            FeatureDefinition (Interval a) a (Int, Maybe b)
+countOfHospitalEventsDef = defineFEF Excluded
+      ( \i es ->
             ((\x -> (length x, duration <$> safeLast x))
             .filterNotDisjoint i                    -- filter to intervals not disjoint from interval
             .combineIntervals                       -- combine overlapping intervals
             .intervals                              -- extract intervals
             .makeConceptsFilter ["wasHospitalized"])  -- filter to only antibiotics events
             es
-
-so :: (IntervalAlgebraic Interval a)=> ComparativePredicateOf (Interval a)
-so = unionPredicates [startedBy, overlappedBy]
+      )
 
 -- | time of distcontinuation of antibiotics
 --   and time from start of follow up
 --   This needs to be generalized as Nothing could either indicate they didn't 
 --   discontinue or that they simply got no antibiotics records.
-discontinuation :: (IntervalSizeable a b, IntervalCombinable a) =>
-     Feature (Interval a)
-  -> Events a
-  -> Feature (Maybe (a, b))
---   -> Feature (Maybe (Interval a))
-discontinuation (Left _) _    = Left Excluded
-discontinuation (Right i) es  = Right $
-      (\x -> Just (begin x           -- we want the begin of this interval 
-                  , diff (begin x) (begin i)))
+so :: (IntervalAlgebraic Interval a)=> ComparativePredicateOf (Interval a)
+so = unionPredicates [startedBy, overlappedBy]
+
+discontinuationDef :: (IntervalSizeable a b
+                      , IntervalAlgebraic Interval a) =>
+                      FeatureDefinition (Interval a) a (Maybe (a, b))
+discontinuationDef = defineFEF Excluded
+      ( \i es ->
+          (\x -> Just (begin x       -- we want the begin of this interval 
+                 , diff (begin x) (begin i)))
       =<< safeHead                   -- if there are any gaps the first one is the first discontinuation
       =<< gapsWithin i               -- find gaps to intervals clipped to i
       =<< (nothingIfNone (so i)      -- if none of the intervals start or overlap 
@@ -202,7 +150,7 @@ discontinuation (Right i) es  = Right $
          . intervals                 -- extract intervals
          . makeConceptsFilter        -- filter to only antibiotics events
             ["tookAntibiotics"]) es
-
+      )
 
 getUnitFeatures ::
       Events Int
@@ -216,14 +164,14 @@ getUnitFeatures ::
      , Feature (Maybe (Int, Int))
      )
 getUnitFeatures x = (
-    index x
-  , enrolled (bline x) x
-  , hasDuckHistory x
-  , hasMacawHistory x
-  , twoMinorOrOneMajor (bline x) x
-  , timeSinceLastAntibiotics (bline x) x
-  , countOfHospitalEvents (bline x) x
-  , discontinuation (flwup x) x
+    applyEF  indexDef x
+  , applyFEF (enrolledDef 8) (bline x) x
+  , applyFEF duckHxDef (bline x) x
+  , applyFEF macawHxDef (bline x) x
+  , applyFEF twoMinorOrOneMajorDef (bline x) x
+  , applyFEF timeSinceLastAntibioticsDef (bline x) x
+  , applyFEF countOfHospitalEventsDef (bline x) x
+  , applyFEF discontinuationDef (flwup x) x
   )
 
 exampleFeatures1Spec :: Spec
@@ -231,24 +179,24 @@ exampleFeatures1Spec = do
 
     it "getUnitFeatures from exampleEvents1" $
       getUnitFeatures exampleEvents1 `shouldBe`
-      ( Right (unsafeInterval (60 :: Int) (61 ::Int))
-      , Right True
-      , Right (True, Just $ unsafeInterval (51 :: Int) (52 :: Int))
-      , Right (False, Nothing)
-      , Right True
-      , Right $ Just 4
-      , Right (1, Just 8)
-      , Right $ Just (78, 18)
+      ( featureR (beginerval 1 (60 :: Int))
+      , featureR True
+      , featureR (True, Just $ beginerval 1 (51 :: Int))
+      , featureR (False, Nothing)
+      , featureR True
+      , featureR $ Just 4
+      , featureR (1, Just 8)
+      , featureR $ Just (78, 18)
       )
 
     it "getUnitFeatures from exampleEvents2" $
       getUnitFeatures exampleEvents2 `shouldBe`
-      ( Left (Other "No occurrence of Orca bite")
-      , Left Excluded
-      , Left Excluded
-      , Left Excluded
-      , Left Excluded
-      , Left Excluded
-      , Left Excluded
-      , Left Excluded
+      ( featureL (Other "No occurrence of Orca bite")
+      , featureL Excluded
+      , featureL Excluded
+      , featureL Excluded
+      , featureL Excluded
+      , featureL Excluded
+      , featureL Excluded
+      , featureL Excluded
       )