+++ src/Hasklepias.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE OverloadedStrings #-}
 {-|
 Module      : Hasklepias
 Description : Re-exports everything you should need to get up and running with 
@@ -6,12 +5,13 @@ Description : Re-exports everything you should need to get up and running with
 Copyright   : (c) NoviSci, Inc 2020
 License     : BSD3
 Maintainer  : bsaul@novisci.com
-Stability   : experimental
 -}
+
+{-# LANGUAGE OverloadedStrings #-}
+
 module Hasklepias (
       module Hasklepias.Types
     , module Hasklepias.Functions
-    , module Hasklepias.Transformations
     , module IntervalAlgebra
     , module IntervalAlgebra.IntervalUtilities
     , module IntervalAlgebra.PairedInterval
@@ -22,5 +22,4 @@ import IntervalAlgebra.IntervalUtilities
 import IntervalAlgebra.PairedInterval
 import Hasklepias.Types
 import Hasklepias.Functions
-import Hasklepias.Transformations